+flatUpdate
v.flateUpdate takes arrays of new & olds to re-render to a single root element (this eliminates the need for the root to have a wrapper as a child to contain multiple virtual dom nodes).

flatUpdate also takes a start & length argument to re-render specific nodes (thus improving performance if some nodes are static while balancing ease of use).